// website/src/app/bulletin/models/bulletinWorkflow.ts
export class BulletinWorkflow {
    workflowId: number;
    submitterUserName: string;
    reviewerUserName: string;
    reviewStatus: string;
    reviewComments:string;
    submissionTime: Date;
    reviewTime: Date;
    bulletinTitle: string;
    bulletinDesc: string;
    bulletinPosterLoc: string;
    clubName: string;
    bulletinStartTime: Date;
    bulletinEndTime: Date;
}